requirements.txt
mysql-connector-python
numpy
librosa
pydub
